+++ modal/functions.py
@@ -49,7 +49,6 @@ from ._utils.function_utils import (
     _create_input,
     _process_result,
     _stream_function_call_data,
-    get_referred_objects,
     is_async,
 )
 from ._utils.grpc_utils import retry_transient_errors
@@ -71,7 +70,7 @@ from .gpu import GPU_T, parse_gpu_config
 from .image import _Image
 from .mount import _get_client_mount, _Mount, get_auto_mounts
 from .network_file_system import _NetworkFileSystem, network_file_system_mount_protos
-from .object import Object, _get_environment_name, _Object, live_method, live_method_gen
+from .object import _get_environment_name, _Object, live_method, live_method_gen
 from .parallel_map import (
     _for_each_async,
     _for_each_sync,
@@ -236,7 +235,7 @@ class FunctionStats:
 
 def _parse_retries(
     retries: Optional[Union[int, Retries]],
-    object_name: str,
+    source: str = "",
 ) -> Optional[api_pb2.FunctionRetryPolicy]:
     if isinstance(retries, int):
         return Retries(
@@ -249,9 +248,9 @@ def _parse_retries(
     elif retries is None:
         return None
     else:
-        raise InvalidError(
-            f"{object_name} retries must be an integer or instance of modal.Retries. Found: {type(retries)}"
-        )
+        extra = f" on {source}" if source else ""
+        msg = f"Retries parameter must be an integer or instance of modal.Retries. Found: {type(retries)}{extra}."
+        raise InvalidError(msg)
 
 
 @dataclass
@@ -289,10 +288,9 @@ class _Function(_Object, type_prefix="fu"):
     _app: Optional["modal.app._App"] = None
     _obj: Optional["modal.cls._Obj"] = None  # only set for InstanceServiceFunctions and bound instance methods
     _web_url: Optional[str]
-    _is_remote_cls_method: bool = False  # TODO(erikbern): deprecated
     _function_name: Optional[str]
     _is_method: bool
-    _spec: _FunctionSpec
+    _spec: Optional[_FunctionSpec] = None
     _tag: str
     _raw_f: Callable[..., Any]
     _build_args: dict
@@ -395,8 +393,6 @@ class _Function(_Object, type_prefix="fu"):
         fun._all_mounts = class_service_function._all_mounts
         fun._spec = class_service_function._spec
         fun._is_method = True
-        # TODO: set more attributes?
-
         return fun
 
     def _bind_instance_method(self, class_bound_method: "_Function"):
@@ -407,28 +403,39 @@ class _Function(_Object, type_prefix="fu"):
         it does it forward invocations to the underlying instance_service_function with the specified method,
         and we don't support web_config for parameterized methods at the moment.
         """
+        # TODO(elias): refactor to not use `_from_loader()` as a crutch for lazy-loading the
+        #   underlying instance_service_function. It's currently used in order to take advantage
+        #   of resolver logic and get "chained" resolution of lazy loads, even though this thin
+        #   object itself doesn't need any "loading"
         instance_service_function = self
         assert instance_service_function._obj
         method_name = class_bound_method._use_method_name
         full_function_name = f"{class_bound_method._function_name}[parameterized]"
 
-        def hydrate_from_instance_function(obj):
-            obj._hydrate_from_other(instance_service_function)
-            obj._obj = instance_service_function._obj
-            obj._web_url = class_bound_method._web_url  # TODO: this shouldn't be set when actual parameters are used
-            obj._function_name = full_function_name
-            obj._is_generator = class_bound_method._is_generator
-            obj._use_method_name = method_name
-            obj._use_function_id = instance_service_function.object_id
+        def hydrate_from_instance_service_function(method_placeholder_fun):
+            method_placeholder_fun._hydrate_from_other(instance_service_function)
+            method_placeholder_fun._obj = instance_service_function._obj
+            method_placeholder_fun._web_url = (
+                class_bound_method._web_url
+            )  # TODO: this shouldn't be set when actual parameters are used
+            method_placeholder_fun._function_name = full_function_name
+            method_placeholder_fun._is_generator = class_bound_method._is_generator
+            method_placeholder_fun._use_method_name = method_name
+            method_placeholder_fun._use_function_id = instance_service_function.object_id
+            method_placeholder_fun._is_method = True
 
         async def _load(fun: "_Function", resolver: Resolver, existing_object_id: Optional[str]):
             # there is currently no actual loading logic executed to create each method on
             # the *parameterized* instance of a class - it uses the parameter-bound service-function
             # for the instance. This load method just makes sure to set all attributes after the
             # `instance_service_function` has been loaded (it's in the `_deps`)
-            hydrate_from_instance_function(fun)
+            hydrate_from_instance_service_function(fun)
 
         def _deps():
+            if instance_service_function.is_hydrated:
+                # without this check, the common instance_service_function will be reloaded by all methods
+                # TODO(elias): Investigate if we can fix this multi-loader in the resolver - feels like a bug?
+                return []
             return [instance_service_function]
 
         rep = f"Method({full_function_name})"
@@ -439,13 +446,9 @@ class _Function(_Object, type_prefix="fu"):
             deps=_deps,
             hydrate_lazily=True,
         )
-        if instance_service_function._can_use_base_function and instance_service_function.is_hydrated:
-            # Eager hydration (skip load) if both the instance service function is already loaded
-            # This should only happen when default arguments are used to the constructor, since
-            # that will trigger similar eager hydration of the instance service function.
-            # In other cases, the instance service function will have to be lazy-loaded, as a dependency
-            # of this "fake" Function.
-            hydrate_from_instance_function(fun)
+        if instance_service_function.is_hydrated:
+            # Eager hydration (skip load) if the instance service function is already loaded
+            hydrate_from_instance_service_function(fun)
 
         fun._info = class_bound_method._info
         fun._obj = instance_service_function._obj
@@ -453,6 +456,7 @@ class _Function(_Object, type_prefix="fu"):
         fun._parent = instance_service_function._parent
         fun._app = class_bound_method._app
         fun._all_mounts = class_bound_method._all_mounts  # TODO: only used for mount-watching/modal serve
+        fun._spec = class_bound_method._spec
         return fun
 
     @staticmethod
@@ -548,8 +552,19 @@ class _Function(_Object, type_prefix="fu"):
             all_mounts = []
 
         retry_policy = _parse_retries(
-            retries, f"Function {info.get_tag()}" if info.raw_f else f"Class {info.get_tag()}"
+            retries, f"Function '{info.get_tag()}'" if info.raw_f else f"Class '{info.get_tag()}'"
         )
+
+        if webhook_config is not None and retry_policy is not None:
+            raise InvalidError(
+                "Web endpoints do not support retries.",
+            )
+
+        if is_generator and retry_policy is not None:
+            deprecation_warning(
+                (2024, 6, 25),
+                "Retries for generator functions are deprecated and will soon be removed.",
+            )
 
         gpu_config = parse_gpu_config(gpu)
 
@@ -669,22 +684,6 @@ class _Function(_Object, type_prefix="fu"):
                 if cloud_bucket_mount.secret:
                     deps.append(cloud_bucket_mount.secret)
 
-            # Add implicit dependencies from the function's code
-            if info.raw_f:
-                # TODO(elias): Remove this branch since we shouldn't need closure vars inspection anymore?
-                objs: list[Object] = get_referred_objects(info.raw_f)
-                _objs: list[_Object] = synchronizer._translate_in(objs)  # type: ignore
-            else:
-                _objs = []
-
-            if info.cls:
-                from .partial_function import _find_callables_for_cls, _PartialFunctionFlags
-
-                for method_callable in _find_callables_for_cls(info.cls, _PartialFunctionFlags.all()).values():
-                    method_objs: list[Object] = get_referred_objects(method_callable)
-                    _objs += synchronizer._translate_in(method_objs)  # type: ignore
-
-            deps += _objs
             return deps
 
         async def _preload(self: _Function, resolver: Resolver, existing_object_id: Optional[str]):
@@ -801,7 +800,7 @@ class _Function(_Object, type_prefix="fu"):
                     allow_concurrent_inputs=allow_concurrent_inputs or 0,
                     worker_id=config.get("worker_id"),
                     is_auto_snapshot=is_auto_snapshot,
-                    is_method=bool(info.cls),
+                    is_method=bool(info.cls) and not info.is_service_class(),
                     checkpointing_enabled=enable_memory_snapshot,
                     is_checkpointing_function=False,
                     object_dependencies=object_dependencies,
@@ -899,6 +898,7 @@ class _Function(_Object, type_prefix="fu"):
                 environment_name=environment_name
                 or "",  # TODO: investigate shouldn't environment name always be specified here?
             )
+
             response = await retry_transient_errors(self._parent._client.stub.FunctionBindParams, req)
             self._hydrate(response.bound_function_id, self._parent._client, response.handle_metadata)
 
@@ -911,11 +911,8 @@ class _Function(_Object, type_prefix="fu"):
             # if the instance didn't use explicit constructor arguments
             fun._hydrate_from_other(self)
 
-        fun._is_remote_cls_method = True  # TODO(erikbern): deprecated
         fun._info = self._info
         fun._obj = obj
-        fun._is_generator = self._is_generator  # TODO(elias): remove - this doesn't apply to "service functions"
-        fun._is_method = False
         fun._parent = self
         return fun
 
@@ -940,9 +937,9 @@ class _Function(_Object, type_prefix="fu"):
             raise InvalidError(
                 textwrap.dedent(
                     """
-                The `.keep_warm()` method can no longer be used on Modal class methods.
-                For classes, all methods now share the same set of containers.
-                Use class_instance.keep_warm(...) instead for classes.
+                The `.keep_warm()` method can not be used on Modal class *methods* deployed using Modal >v0.63.
+
+                Call `.keep_warm()` on the class *instance* instead.
             """
                 )
             )
@@ -1038,6 +1035,7 @@ class _Function(_Object, type_prefix="fu"):
     @property
     def spec(self) -> _FunctionSpec:
         """mdmd:hidden"""
+        assert self._spec
         return self._spec
 
     def get_build_def(self) -> str:
@@ -1087,6 +1085,7 @@ class _Function(_Object, type_prefix="fu"):
             web_url=self._web_url or "",
             use_method_name=self._use_method_name,
             use_function_id=self._use_function_id,
+            is_method=self._is_method,
         )
 
     def _set_mute_cancellation(self, value: bool = True):
@@ -1219,15 +1218,12 @@ class _Function(_Object, type_prefix="fu"):
         else:
             await self._call_function(args, kwargs)
 
-    def _get_is_remote_cls_method(self):
-        return self._is_remote_cls_method
-
     def _get_info(self) -> FunctionInfo:
         if not self._info:
             raise ExecutionError("Can't get info for a function that isn't locally defined")
         return self._info
 
-    def _get_obj(self) -> Optional[Any]:
+    def _get_obj(self) -> Optional["modal.cls._Obj"]:
         if not self._is_method:
             return None
         elif not self._obj:
@@ -1247,6 +1243,12 @@ class _Function(_Object, type_prefix="fu"):
         # TODO(erikbern): it would be nice to remove the nowrap thing, but right now that would cause
         # "user code" to run on the synchronicity thread, which seems bad
         info = self._get_info()
+
+        if is_local() and self.spec.volumes or self.spec.network_file_systems:
+            warnings.warn(
+                f"The {info.function_name} function is executing locally "
+                + "and will not have access to the mounted Volume or NetworkFileSystem data"
+            )
         if not info or not info.raw_f:
             msg = (
                 "The definition for this function is missing so it is not possible to invoke it locally. "
@@ -1254,16 +1256,16 @@ class _Function(_Object, type_prefix="fu"):
             )
             raise ExecutionError(msg)
 
-        obj = self._get_obj()
+        obj: Optional["modal.cls._Obj"] = self._get_obj()
 
         if not obj:
             fun = info.raw_f
             return fun(*args, **kwargs)
         else:
             # This is a method on a class, so bind the self to the function
-            local_obj = obj.get_local_obj()
+            user_cls_instance = obj._get_user_cls_instance()
 
-            fun = info.raw_f.__get__(local_obj)
+            fun = info.raw_f.__get__(user_cls_instance)
 
             if is_async(info.raw_f):
                 # We want to run __aenter__ and fun in the same coroutine