+++ test/conftest.py
@@ -139,6 +139,7 @@ class MockClientServicer(api_grpc.ModalClientBase):
         self.app_functions: Dict[str, api_pb2.Function] = {}
         self.bound_functions: Dict[Tuple[str, bytes], str] = {}
         self.function_params: Dict[str, Tuple[Tuple, Dict[str, Any]]] = {}
+        self.function_options: Dict[str, api_pb2.FunctionOptions] = {}
         self.fcidx = 0
 
         self.function_serialized = None
@@ -554,9 +555,19 @@ class MockClientServicer(api_grpc.ModalClientBase):
         for k, v in self.dicts[request.dict_id].items():
             await stream.send_message(api_pb2.DictEntry(key=k, value=v))
 
+    ### Environment
+
+    async def EnvironmentCreate(self, stream):
+        await stream.send_message(Empty())
+
+    async def EnvironmentUpdate(self, stream):
+        await stream.send_message(api_pb2.EnvironmentListItem())
+
     ### Function
 
     async def FunctionBindParams(self, stream):
+        from modal._serialization import deserialize
+
         request: api_pb2.FunctionBindParamsRequest = await stream.recv_message()
         assert request.function_id
         assert request.serialized_params
@@ -573,9 +584,8 @@ class MockClientServicer(api_grpc.ModalClientBase):
         bound_func.CopyFrom(base_function)
         self.app_functions[function_id] = bound_func
         self.bound_functions[(request.function_id, request.serialized_params)] = function_id
-        from modal._serialization import deserialize
-
         self.function_params[function_id] = deserialize(request.serialized_params, None)
+        self.function_options[function_id] = request.function_options
 
         await stream.send_message(
             api_pb2.FunctionBindParamsResponse(
@@ -1287,9 +1297,9 @@ class MockClientServicer(api_grpc.ModalClientBase):
 
     async def VolumeGetFile(self, stream):
         req = await stream.recv_message()
-        if req.path.decode("utf-8") not in self.volume_files[req.volume_id]:
+        if req.path not in self.volume_files[req.volume_id]:
             raise GRPCError(Status.NOT_FOUND, "File not found")
-        vol_file = self.volume_files[req.volume_id][req.path.decode("utf-8")]
+        vol_file = self.volume_files[req.volume_id][req.path]
         if vol_file.data_blob_id:
             await stream.send_message(api_pb2.VolumeGetFileResponse(data_blob_id=vol_file.data_blob_id))
         else:
@@ -1305,9 +1315,9 @@ class MockClientServicer(api_grpc.ModalClientBase):
 
     async def VolumeRemoveFile(self, stream):
         req = await stream.recv_message()
-        if req.path.decode("utf-8") not in self.volume_files[req.volume_id]:
+        if req.path not in self.volume_files[req.volume_id]:
             raise GRPCError(Status.INVALID_ARGUMENT, "File not found")
-        del self.volume_files[req.volume_id][req.path.decode("utf-8")]
+        del self.volume_files[req.volume_id][req.path]
         await stream.send_message(Empty())
 
     async def VolumeListFiles(self, stream):
@@ -1352,26 +1362,23 @@ class MockClientServicer(api_grpc.ModalClientBase):
     async def VolumeCopyFiles(self, stream):
         req = await stream.recv_message()
         for src_path in req.src_paths:
-            if src_path.decode("utf-8") not in self.volume_files[req.volume_id]:
+            if src_path not in self.volume_files[req.volume_id]:
                 raise GRPCError(Status.NOT_FOUND, f"Source file not found: {src_path}")
-            src_file = self.volume_files[req.volume_id][src_path.decode("utf-8")]
+            src_file = self.volume_files[req.volume_id][src_path]
             if len(req.src_paths) > 1:
                 # check to make sure dst is a directory
-                if (
-                    req.dst_path.decode("utf-8").endswith(("/", "\\"))
-                    or not os.path.splitext(os.path.basename(req.dst_path))[1]
-                ):
+                if req.dst_path.endswith(("/", "\\")) or not os.path.splitext(os.path.basename(req.dst_path))[1]:
                     dst_path = os.path.join(req.dst_path, os.path.basename(src_path))
                 else:
                     raise GRPCError(Status.INVALID_ARGUMENT, f"{dst_path} is not a directory.")
             else:
                 dst_path = req.dst_path
-            self.volume_files[req.volume_id][dst_path.decode("utf-8")] = src_file
+            self.volume_files[req.volume_id][dst_path] = src_file
         await stream.send_message(Empty())
 
 
-@pytest_asyncio.fixture
-async def blob_server():
+@pytest.fixture
+def blob_server():
     blobs = {}
     blob_parts: Dict[str, Dict[int, bytes]] = defaultdict(dict)
 
@@ -1415,8 +1422,30 @@ async def blob_server():
     app.add_routes([aiohttp.web.get("/download", download)])
     app.add_routes([aiohttp.web.post("/complete_multipart", complete_multipart)])
 
-    async with run_temporary_http_server(app) as host:
-        yield host, blobs
+    started = threading.Event()
+    stop_server = threading.Event()
+
+    host = None
+
+    def run_server_other_thread():
+        loop = asyncio.new_event_loop()
+
+        async def async_main():
+            nonlocal host
+            async with run_temporary_http_server(app) as _host:
+                host = _host
+                started.set()
+                await loop.run_in_executor(None, stop_server.wait)
+
+        loop.run_until_complete(async_main())
+
+    # run server on separate thread to not lock up the server event loop in case of blocking calls in tests
+    thread = threading.Thread(target=run_server_other_thread)
+    thread.start()
+    started.wait()
+    yield host, blobs
+    stop_server.set()
+    thread.join()
 
 
 @pytest_asyncio.fixture(scope="function")