+++ test/supports/functions.py
@@ -7,8 +7,6 @@ from typing import List
 
 from modal import (
     App,
-    Image,
-    Volume,
     asgi_app,
     build,
     current_function_call_id,
@@ -261,13 +259,19 @@ class LifecycleCls:
         await asyncio.sleep(self.async_exit_duration)
 
     @method()
+    def local(self):
+        self.events.append("local")
+
+    @method()
     def f_sync(self):
         self.events.append("f_sync")
+        self.local.local()
         return self.events
 
     @method()
     async def f_async(self):
         self.events.append("f_async")
+        self.local.local()
         return self.events
 
     @method()
@@ -359,20 +363,6 @@ def function_calling_method(x, y, z):
     return obj.f.remote(z)
 
 
-image = Image.debian_slim().pip_install("xyz")
-other_image = Image.debian_slim().pip_install("abc")
-volume = Volume.from_name("vol", create_if_missing=True)
-other_volume = Volume.from_name("other-vol", create_if_missing=True)
-
-
-@app.function(image=image, volumes={"/tmp/xyz": volume})
-def check_dep_hydration(x):
-    assert image.is_hydrated
-    assert other_image.is_hydrated
-    assert volume.is_hydrated
-    assert other_volume.is_hydrated
-
-
 @app.cls()
 class BuildCls:
     def __init__(self):
@@ -443,3 +433,9 @@ def sandbox_f(x):
 @app.function()
 def is_local_f(x):
     return is_local()
+
+
+@app.function()
+def raise_large_unicode_exception():
+    byte_str = (b"k" * 120_000_000) + b"\x99"
+    byte_str.decode("utf-8")